Ataque Especial - versão inicial
funcionando disparo do projétil, porém o dano está absurdo, preciso descobrir qual o problema na colisão pra que não mate o chefe com apenas 1 disparo -  na reta final darei os retoques

+++ RetroGame/Assets/scripts/2D/Level03/Boss/BossCombat.cs
@@ -8,7 +8,7 @@ public class BossCombat : MonoBehaviour
 {
     [SerializeField] Transform punchAttack;
     public float punchRange = 0.5f;
-    private int punchDamage = 10;
+    private float punchDamage = 10;
     public float repulsionForce = 100f;
     public LayerMask playerLayer;
